+++ src/Facility.Core/ServiceDto.cs
@@ -33,7 +33,7 @@ public abstract class ServiceDto
 	/// </summary>
 	protected virtual JsonServiceSerializer JsonSerializer => JsonServiceSerializer.Legacy;
 
-	internal ServiceObject ToServiceObject() => JsonSerializer.ToServiceObject(this);
+	internal JsonServiceSerializer GetJsonSerializer() => JsonSerializer;
 }
 
 /// <summary>
@@ -51,5 +51,10 @@ public abstract class ServiceDto<T> : ServiceDto
 	/// <summary>
 	/// Determines if two DTOs are equivalent.
 	/// </summary>
-	public abstract bool IsEquivalentTo(T? other);
+	public virtual bool IsEquivalentTo(T? other) => JsonSerializer.AreEquivalent(this, other);
+
+	/// <summary>
+	/// Deep clones the DTO.
+	/// </summary>
+	public virtual T DeepClone() => JsonSerializer.Clone((T) this);
 }