+++ tests/Facility.Benchmarks/SerializerInfo.cs
@@ -2,7 +2,7 @@ using Facility.Core;
 
 namespace Facility.Benchmarks;
 
-public record SerializerInfo(ServiceSerializer ServiceSerializer)
+internal sealed record SerializerInfo(ServiceSerializer ServiceSerializer)
 {
 	public override string ToString() => ServiceSerializer.GetType().Name[0..^"ServiceSerializer".Length];
 }