+++ tests/Facility.Core.UnitTests/Http/AuthorizationClientAspectTests.cs
@@ -1,3 +1,6 @@
+#if NET472
+using System.Net.Http;
+#endif
 using Facility.Core.Http;
 using FluentAssertions;
 using NUnit.Framework;
@@ -6,7 +9,7 @@ using NUnit.Framework;
 
 namespace Facility.Core.UnitTests.Http;
 
-public class AuthorizationClientAspectTests
+internal sealed class AuthorizationClientAspectTests
 {
 	[Test]
 	public async Task AuthorizationClientAspectWorks()