+++ tests/Facility.Core.UnitTests/TestDto.cs
@@ -1,9 +1,11 @@
+using System.Diagnostics.CodeAnalysis;
 using MessagePack;
 
 namespace Facility.Core.UnitTests;
 
 [MessagePackObject]
-public class TestDto : ServiceDto<TestDto>
+[SuppressMessage("Maintainability", "CA1515:Consider making public types internal", Justification = "Must be public for MessagePack.")]
+public sealed class TestDto : SystemTextJsonServiceDto<TestDto>
 {
 	[Key(0)]
 	public int? Id { get; set; }
@@ -18,13 +20,4 @@ public class TestDto : ServiceDto<TestDto>
 	[Newtonsoft.Json.JsonProperty(DefaultValueHandling = Newtonsoft.Json.DefaultValueHandling.Ignore, NullValueHandling = Newtonsoft.Json.NullValueHandling.Include), ServiceNullableDefaultValue(typeof(ServiceNullable<bool?>))]
 	[System.Text.Json.Serialization.JsonIgnore(Condition = System.Text.Json.Serialization.JsonIgnoreCondition.WhenWritingDefault)]
 	public ServiceNullable<bool?> Ternary { get; set; }
-
-	public override bool IsEquivalentTo(TestDto? other)
-	{
-		return other != null &&
-			other.Id == Id &&
-			other.Name == Name &&
-			ServiceDataUtility.AreEquivalentArrays(other.Children, Children, ServiceDataUtility.AreEquivalentDtos) &&
-			other.Ternary == Ternary;
-	}
 }